fix(Schemas): :bug: Correct schema filenames
