+++ app/schemas/satisfaction.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+
+class SatisfactionRequest(BaseModel):
+    satisfied: bool
